글 작성 기능 및 목록 불러오기 기능 일부 작성

+++ src/main/java/com/StarJ/Social/Controllers/UserController.java
@@ -2,8 +2,8 @@ package com.StarJ.Social.Controllers;
 
 import com.StarJ.Social.DTOs.UserRequestDTO;
 import com.StarJ.Social.DTOs.UserResponseDTO;
-import com.StarJ.Social.Domains.SiteUser;
-import com.StarJ.Social.Securities.JWT.JwtTokenProvider;
+import com.StarJ.Social.Domains.LocalFile;
+import com.StarJ.Social.Service.AuthService;
 import com.StarJ.Social.Service.LocalFileService;
 import com.StarJ.Social.Service.UserService;
 import lombok.RequiredArgsConstructor;
@@ -15,53 +15,54 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/api/user")
 @RequiredArgsConstructor
 public class UserController {
-
+    private final AuthService authService;
     private final UserService userService;
-    private final JwtTokenProvider jwtTokenProvider;
-    private final LocalFileService localFileService;
+
     @PostMapping("/signup")
     public ResponseEntity<?> singUp(@RequestBody UserRequestDTO requestDto) {
         this.userService.signup(requestDto);
         return ResponseEntity.status(HttpStatus.OK).body(null);
     }
+
     @GetMapping("/data")
     public ResponseEntity<?> getData(@RequestHeader("Username") String username) {
         try {
             UserResponseDTO userResponseDto = this.userService.findById(username);
-            userResponseDto.setProfileImage(localFileService.getProfileImage(username));
             return ResponseEntity.status(HttpStatus.OK).body(userResponseDto);
         } catch (IllegalArgumentException ex) {
             return ResponseEntity.status(HttpStatus.OK).body(null);
         }
     }
 
-    @GetMapping("/user")
+    @GetMapping
     public ResponseEntity<?> findUser(@RequestHeader("Authorization") String accessToken) {
-        if (accessToken != null && accessToken.length() > 7) {
-            String token = accessToken.substring(7);
-            if (this.jwtTokenProvider.validateToken(token)) {
-                String username = this.jwtTokenProvider.getUsernameFromToken(token);
-                UserResponseDTO userResponseDto = this.userService.findById(username);
-                userResponseDto.setProfileImage(localFileService.getProfileImage(username));
-                return ResponseEntity.status(HttpStatus.OK).body(userResponseDto);
-            } else
-                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        AuthService.TokenReturnClass tokenReturnClass = authService.checkToken(accessToken);
+        if (tokenReturnClass.isOK()) {
+            String username = tokenReturnClass.username();
+            UserResponseDTO userResponseDto = this.userService.findById(username);
+            return ResponseEntity.status(HttpStatus.OK).body(userResponseDto);
         } else
-            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+            return tokenReturnClass.getResponseEntity();
     }
 
-    @PutMapping("/user")
+    @PutMapping
     public ResponseEntity<?> updateUser(@RequestHeader("Authorization") String accessToken,
                                         @RequestBody UserRequestDTO requestDto) {
-        String username = this.jwtTokenProvider.getUsernameFromToken(accessToken.substring(7));
-        this.userService.update(username, requestDto);
-        return ResponseEntity.status(HttpStatus.OK).body(null);
+        AuthService.TokenReturnClass tokenReturnClass = authService.checkToken(accessToken);
+        if (tokenReturnClass.isOK()) {
+            String username = tokenReturnClass.username();
+            this.userService.update(username, requestDto);
+        }
+        return tokenReturnClass.getResponseEntity();
     }
 
-    @DeleteMapping("/user")
+    @DeleteMapping
     public ResponseEntity<?> deleteUser(@RequestHeader("Authorization") String accessToken) {
-        String username = this.jwtTokenProvider.getUsernameFromToken(accessToken.substring(7));
-        this.userService.delete(username);
-        return ResponseEntity.status(HttpStatus.OK).body(null);
+        AuthService.TokenReturnClass tokenReturnClass = authService.checkToken(accessToken);
+        if (tokenReturnClass.isOK()) {
+            String username = tokenReturnClass.username();
+            this.userService.delete(username);
+        }
+        return tokenReturnClass.getResponseEntity();
     }
 }
